refactor: split orchestrator extension into logical modules
Split 2189-line index.ts into 12 focused modules:
- index.ts (37 lines) — main entry, imports and wires all modules
- utils.ts — terminalNotify, isRunningInContainer, getPiInvocation
- git-helpers.ts — runGit, getCurrentBranch, isBranchMerged, etc.
- enforcement.ts — tool_call handler (python/git/dangerous commands)
- subagent-tool.ts — subagent tool + runSingleAgent + render logic
- async-agents.ts — async infrastructure (state, poller, watcher, spawner)
- ask-user.ts — ask_user tool
- btw.ts — /btw command
- status-line.ts — git status, container indicator, notifications
- rules.ts — before_agent_start rule injection
- session-validation.ts — session start tool checks
